set sw=0 sts=-1 modeline (and remove laststatus nobackup)
often happens that someone has ts=8 in the modeline, so with my previous
settings this would result in the GNU-style ts=8 sw=4.

sts is mostly irrelevant (vim-sensible sets smarttab, which makes tabs
at the start of the line use sw instead, and that's the only place where
tabs should be placed anyway). Nontheless, it's better to have it
default to sw than to ts.

laststatus=2 is already in vim-sensible, and nobackup seems to be the
default (perhaps it wasn't on some other system).